Split Trakt client into its own file, add more typing and search method
